Added a example program to get weather info.
The program "getWeather.py" will download weather information from the Weather Underground Mars Hill station (KAZFLAGS163) and then make a plot of the temperature and humidity.  When the program is run it expects a subdirectory called "data" where it puts CSV files containing the weather data.  The program gets data for Nov. 4-5, 2018 but you can change that if you like by editing the program.